V3/v3/pyproject.toml
[project]
name = "phishing-agent"
version = "3.0.0"
description = "Agent de detection de phishing via Microsoft Graph API"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.15.0",
    "msgraph-sdk>=1.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "typer[all]>=0.9.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "apscheduler>=3.10.4",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "httpx>=0.26.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
ml = [
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "joblib>=1.3.0",
]

[project.scripts]
phishing-agent = "src.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
